update: PaginationをPostListと分離してpage.tsxでそれぞれ読めるように修正
