+++ src/components/Page/Archive/Pagination/constant.ts
@@ -1,0 +1,3 @@
+export const QUERY_PAGE_KEY = 'p';
+export const ITEMS_PER_PAGE = 5;
+export const DOTS = '...';
