+++ src/components/Page/Archive/Pagination/index.ts
@@ -1,2 +1,2 @@
+export { PostList } from './PostList';
 export { Pagination } from './Pagination';
-export { usePagination, DOTS } from './usePagination';
